requirements.txt
streamlit
pandas
python-pptx
openai
